Add deselection functionality
Now you can deselect cells in the same you use to select cells if you
hold the ctrl button.
This commit also fixes a bug that was introduced when trying to optimise
the mousemove event listener.
The select/deselect column/row function were removed and their
functionality was inntegrated in the relevant parts in the programme to
enhance the performance.